LArSoft Supera module implemented. It's literally a copy from LArLite Supera.cxx except for changing include path and use of art handles (still 99% are identical copy) --kazu

+++ core/larfmwk_shared/MCParticleTree.inl
@@ -23,7 +23,7 @@ namespace larcv {
 
 	auto const& mctruth = mctruth_v[mctruth_index];
 
-	for(size_t i=0; i<mctruth.NParticles(); ++i) {
+	for(size_t i=0; i<(size_t)(mctruth.NParticles()); ++i) {
 
 	  auto const& mcp = mctruth.GetParticle(i);
